# Use stackalloc to reduce allocations on NewUlid